+++ src/c/ncm_mutable_song.h
@@ -28,7 +28,7 @@ typedef struct NcmMutableSong {
     int32 name_len;
     int32 new_name_len;
 
-    int64 mtime;
+    int32 mtime;
     int32 duration;
     bool is_from_database;
 
@@ -79,8 +79,8 @@ bool ncm_mutable_song_get_new_name(NcmMutableSong *song, NcmStringView *view);
 
 void ncm_mutable_song_set_duration(NcmMutableSong *song, int32 duration);
 int32 ncm_mutable_song_duration(NcmMutableSong *song);
-void ncm_mutable_song_set_mtime(NcmMutableSong *song, int64 mtime);
-int64 ncm_mutable_song_mtime(NcmMutableSong *song);
+void ncm_mutable_song_set_mtime(NcmMutableSong *song, int32 mtime);
+int32 ncm_mutable_song_mtime(NcmMutableSong *song);
 
 bool ncm_mutable_song_is_modified(NcmMutableSong *song);
 void ncm_mutable_song_clear_modifications(NcmMutableSong *song);
