+++ src/c/ncm_song.c
@@ -692,10 +692,10 @@ ncm_song_format_track_number(char *buffer, int32 buffer_cap,
 }
 
 int32
-ncm_song_show_time(int64 length, char *buffer, int32 buffer_cap) {
-    int64 hours;
-    int64 minutes;
-    int64 seconds;
+ncm_song_show_time(int32 length, char *buffer, int32 buffer_cap) {
+    int32 hours;
+    int32 minutes;
+    int32 seconds;
     int32 result;
 
     if ((buffer == NULL) || (buffer_cap <= 0)) {
